extra_codes/face_enhancer/requirements.txt
# Python 3.8+ recommended
numpy
opencv-python
Pillow
facenet-pytorch
rembg
gfpgan
scikit-image
tqdm
